libft 완성

+++ libft/srcs/ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: minkyuki <minkyuki@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/08 10:21:17 by minkyuki          #+#    #+#             */
-/*   Updated: 2022/11/08 10:29:12 by minkyuki         ###   ########.fr       */
+/*   Updated: 2022/11/10 14:17:44 by minkyuki         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strrchr(const char *s, int c)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (s[i])
